+++ model_hosting/main.py
@@ -1,0 +1,16 @@
+
+
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+from fast_api import router
+
+app = FastAPI()
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+app.include_router(router)
+
